Make split optionally trim the last element if it is just a trailing newline
Most of our processed data has a trailing newline, but there is no guarantee. By trimming it we make bleualign compatible with either. Note: it was already compatible because there was no checking, but the extra checks I've added to make sure the length of textN and textNtranslated is the same showed the difference. I think it is worth to keep the input validation.